Issue #1023: Extract PageBuilder interface
